Door led is now on pin 2 and blinks on invalid key
There will be a status led on the shield that has the
status blink, and the door led should NOT blink since that
could attract attention.

However the use of a invalid key will make the door led
to blink fast for approx 3s.
And a valid key will turn the led on for the same
time as the lock is open.